+++ poisson_approval/strategies/StrategyThreshold.py
@@ -4,18 +4,31 @@ from poisson_approval.strategies.StrategyTwelve import StrategyTwelve
 from poisson_approval.utils.DictPrintingInOrderIgnoringZeros import DictPrintingInOrderIgnoringZeros
 from poisson_approval.utils.DictPrintingInOrderIgnoringNone import DictPrintingInOrderIgnoringNone
 from poisson_approval.utils.Util import ballot_one, ballot_two, ballot_one_two, ballot_one_three
+from poisson_approval.utils.UtilCache import cached_property
 
 
 # noinspection PyUnresolvedReferences
 class StrategyThreshold(StrategyTwelve):
-    """A strategy profile for a cardinal profile.
+    """A threshold strategy (for a cardinal profile).
+
+    For each ranking, there is a ``threshold`` and a ``ratio_optimistic``. E.g. assume that for ranking ``abc``,
+    the threshold is 0.4 and the ratio is 0.2. It means that:
+
+    * A voter ``abc`` votes for `ab` if their utility for `b` is strictly greater than 0.4,
+    * A voter ``abc`` votes for `a` if their utility for `b` is strictly lower than 0.4,
+    * Voters ``abc`` whose utility for `b` is equal to 0.4 are split: a ratio 0.2 are optimistic, they vote only for
+      `a` (they behave as if the pivot `ab` was very likely); and a ratio 0.8 are pessimistic, they vote for `ab`
+      (they behave as if the pivot `bc` was very likely).
+
+    For a given ranking, the threshold and / or the ratio may be None, which means that they are not specified for
+    this strategy.
 
     Parameters
     ----------
-    d_ranking_threshold : dict
-        Keys are rankings and values are utility thresholds. E.g. ``'abc': 0.4`` means that a voter ``abc`` vote for
-        ``ab`` if her utility for ``b`` is strictly greater than 0.4, and for ``a`` otherwise. A threshold can be
-        ``None``, meaning that the behavior of these voters is not specified in the strategy.
+    d : dict
+        Cf. examples below for the different types of input syntax.
+    ratio_optimistic : Number
+        Cf. examples below for the different types of input syntax.
     profile : Profile, optional
         The "context" in which the strategy is used.
     voting_rule : str
@@ -24,11 +37,25 @@ class StrategyThreshold(StrategyTwelve):
 
     Examples
     --------
-        >>> strategy = StrategyThreshold({'abc': 0.4, 'bac': 0.51, 'cab': 1})
-        >>> strategy
-        StrategyThreshold({'abc': 0.4, 'bac': 0.51, 'cab': 1})
+    The two following examples illustrate different ways to define the same profile. The first possible type of input
+    syntax is a dict that maps a ranking to a tuple (``threshold``, ``ratio_optimistic``). It corresponds exactly to
+    the attribute :attr:`d_ranking_t_threshold_ratio_optimistic`:
+
+        >>> strategy = StrategyThreshold({'abc': (0.4, 0.2), 'bac': (0.51, 0.2), 'cab': (1, 0.2)})
         >>> print(strategy)
-        <abc: utility-dependent (0.4), bac: utility-dependent (0.51), cab: c>
+        <abc: utility-dependent (0.4, 0.2), bac: utility-dependent (0.51, 0.2), cab: c>
+
+    The second  possible type of input syntax is a dict that maps a ranking to a threshold. All rankings have the same
+    ratio, given by the parameter `ratio_optimistic`:
+
+        >>> strategy = StrategyThreshold({'abc': 0.4, 'bac': 0.51, 'cab': 1}, ratio_optimistic=0.2)
+        >>> print(strategy)
+        <abc: utility-dependent (0.4, 0.2), bac: utility-dependent (0.51, 0.2), cab: c>
+
+    Some operations on the strategy:
+
+        >>> strategy
+        StrategyThreshold({'abc': (0.4, 0.2), 'bac': (0.51, 0.2), 'cab': (1, 0.2)})
         >>> strategy.abc
         'utility-dependent'
         >>> strategy.a_bc
@@ -37,13 +64,31 @@ class StrategyThreshold(StrategyTwelve):
         'ab'
         >>> strategy.d_ranking_threshold['abc']
         0.4
+
+    It is possible not to specify the ratios of optimistic voters:
+
+        >>> strategy = StrategyThreshold({'abc': 0.4, 'bac': 0.51})
+        >>> strategy
+        StrategyThreshold({'abc': 0.4, 'bac': 0.51})
+        >>> print(strategy)
+        <abc: utility-dependent (0.4), bac: utility-dependent (0.51)>
+
+    If this strategy is applied to a profile where a positive share of voters have ranking `abc` and a utility 0.2
+    for their middle candidate `b`, it will raise an error because the ratio of optimistic voters is needed in that
+    case.
     """
 
-    def __init__(self, d_ranking_threshold, profile=None, voting_rule=None):
+    def __init__(self, d, ratio_optimistic=None, profile=None, voting_rule=None):
         voting_rule = self._get_voting_rule_(profile, voting_rule)
-        # Populate the dictionary of thresholds
+        # Prepare the dictionaries of thresholds and ratios
         self.d_ranking_threshold = DictPrintingInOrderIgnoringNone({ranking: None for ranking in RANKINGS})
-        self.d_ranking_threshold.update(d_ranking_threshold)
+        self.d_ranking_ratio_optimistic = DictPrintingInOrderIgnoringNone({ranking: None for ranking in RANKINGS})
+        for ranking, value in d.items():
+            if isinstance(value, tuple):
+                self.d_ranking_threshold[ranking], self.d_ranking_ratio_optimistic[ranking] = value
+            else:
+                self.d_ranking_threshold[ranking] = value
+                self.d_ranking_ratio_optimistic[ranking] = ratio_optimistic
         # Prepare the dictionary of ballots
         d_ranking_ballot = DictPrintingInOrderIgnoringZeros()
         for ranking, threshold in self.d_ranking_threshold.items():
@@ -68,6 +113,17 @@ class StrategyThreshold(StrategyTwelve):
         # Call parent class
         super().__init__(d_ranking_ballot=d_ranking_ballot, profile=profile, voting_rule=voting_rule)
 
+    @cached_property
+    def d_ranking_t_threshold_ratio_optimistic(self):
+        """dict : Thresholds and ratios of optimistic voters.
+
+        Key : ranking, e.g. ``'abc'``. Value: tuple (``threshold``, ``ratio_optimistic``).
+        """
+        return DictPrintingInOrderIgnoringNone({
+            ranking: (self.d_ranking_threshold[ranking], self.d_ranking_ratio_optimistic[ranking])
+            for ranking in RANKINGS
+        })
+
     def __eq__(self, other):
         """Equality test.
 
@@ -82,12 +138,13 @@ class StrategyThreshold(StrategyTwelve):
 
         Examples
         --------
-            >>> strategy = StrategyThreshold({'abc': 0.4, 'bac': 0.51, 'cab': 1})
-            >>> strategy == StrategyThreshold({'abc': 0.4, 'bac': 0.51, 'cab': 1})
+            >>> strategy = StrategyThreshold({'abc': (0.4, 0.2), 'bac': (0.51, 0.2), 'cab': (1, 0.2)})
+            >>> strategy == StrategyThreshold({'abc': (0.4, 0.2), 'bac': (0.51, 0.2), 'cab': (1, 0.2)})
             True
         """
         return (isinstance(other, StrategyThreshold)
                 and self.d_ranking_threshold == other.d_ranking_threshold
+                and self.d_ranking_ratio_optimistic == other.d_ranking_ratio_optimistic
                 and self.voting_rule == other.voting_rule)
 
     def isclose(self, other, *args, **kwargs):
@@ -108,26 +165,37 @@ class StrategyThreshold(StrategyTwelve):
 
         Examples
         --------
-            >>> strategy = StrategyThreshold({'abc': 0.4, 'bac': 0.51, 'cab': 1})
-            >>> strategy.isclose(StrategyThreshold({'abc': 0.4, 'bac': 0.51, 'cab': 0.999999999999999999999999}))
+            >>> strategy = StrategyThreshold({'abc': (0.4, 0.2), 'bac': (0.51, 0.2), 'cab': (1, 0.2)})
+            >>> strategy.isclose(StrategyThreshold({
+            ...     'abc': (0.4, 0.2), 'bac': (0.51, 0.1999999999999), 'cab': (0.999999999999, 0.2)}))
             True
         """
         return isinstance(other, StrategyThreshold) and all([
             (threshold is None and other.d_ranking_threshold[ranking] is None)
             or isclose(threshold, other.d_ranking_threshold[ranking], *args, **kwargs)
             for ranking, threshold in self.d_ranking_threshold.items()
+        ]) and all([
+            (ratio_optimistic is None and other.d_ranking_ratio_optimistic[ranking] is None)
+            or isclose(ratio_optimistic, other.d_ranking_ratio_optimistic[ranking], *args, **kwargs)
+            for ranking, ratio_optimistic in self.d_ranking_ratio_optimistic.items()
         ])
 
     def __repr__(self):
-        arguments = repr(self.d_ranking_threshold)
+        d = DictPrintingInOrderIgnoringNone({
+            k: t[0] if t[1] is None else t
+            for k, t in self.d_ranking_t_threshold_ratio_optimistic.items()})
+        arguments = repr(d)
         if self.voting_rule != APPROVAL:
             arguments += ', voting_rule=%r' % self.voting_rule
         return 'StrategyThreshold(%s)' % arguments
 
     def __str__(self):
+        def t_threshold_ratio_to_string(t):
+            return str(t[0]) if t[1] is None else '%s, %s' % t
+
         result = '<' + ', '.join([
             '%s: %s' % (ranking, self.d_ranking_ballot[ranking])
-            + (' (%s)' % self.d_ranking_threshold[ranking]
+            + (' (%s)' % t_threshold_ratio_to_string(self.d_ranking_t_threshold_ratio_optimistic[ranking])
                if self.d_ranking_ballot[ranking] == UTILITY_DEPENDENT else '')
             for ranking in sorted(self.d_ranking_ballot) if self.d_ranking_ballot[ranking]
         ]) + '>'